requirements.txt
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.5.7
colorama==0.4.4
Django==3.2.7
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
executing==0.8.1
gunicorn==20.1.0
icecream==2.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
Pygments==2.10.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
